Updated document model and added document processing app
