+++ backend/VocabularyManagement/models.py
@@ -1,17 +1,15 @@
 from django.db import models
 from django.contrib.auth.models import User
-from backend.DocumentManagement.models import Document
+from DocumentManagement.models import Document
 
 # Create your models here.
 
 #Vocabulary Management related models/functionality:
-#Document Preprocessing
-    #List of words in Doc
-#Highlight exraction
 #Vocabulary
 #Study Set
 
 
+#Started study set - will expand on in future sprints
 class StudySet(models.Model):
     owner_id = models.ForeignKey(User, on_delete=models.CASCADE)
     generated_by = models.ForeignKey(Document, on_delete=models.CASCADE, null=True)
